Implemented Phase 1 of the Weak Points Mechanic
The way it's implemented should make it capable of working out-of-the-box. Nothing should happen if the hitboxes are left empty, though the furniture will only take normal damage if the hitboxes aren't set. Otherwise the way it's supposed to be implemented is to have multiple collision instances on the furniture at once, each manually assigned on the Shootable, which has an array of Hitbox structs. For now it only has the collider it's associated with, and the damage multiplier. If the AI is expanded upon more, we may add some more variables to play with that might be able to interact with the AI through the Shootable script.