Implemented a Limping Mechanic
Expanded upon Shootable to include a GetHealth() method which returns an array of two integers, the first integer being the currentHealth within Shootable, and the second integer being maxHealth pulled directly from the FurnitureSO.
Expanded upon WanderAI's Update() method to also call upon GetHealth from Shootable, in order to multiply the agent's original speed from FurnitureSO by a basic graph formula of y = 1 - x^3, where x is the difference in maxHealth and currentHealth, divided by maxHealth. The formula is set out in a way where speed reduction only starts to become significant past half maxHealth. For example, at 50% health speed is down to 87.5%, 40% health speed is down to 78.4%, 30% health speed is down to 65.7%, 20% health speed is down to 48.8%, 10% health speed is down to 27.1%, etc. The math formula can be adjusted accordingly, whether to raise the power or lower it to sharpen or loosen up the exponential decrease in speed compared with health. Alternatively the formula can be inverted so that minor damage can cause a significant speed decrease with diminishing speed loss at lower health. The final formula to decide upon limping speed is up for discussion.