Furniture Sensing Part 1
Changes:

FurnitureSO now has new properties:
- float height: Y offset that is used for senses. May/may not be used, but has been added just in case.
- bool xray: Determines whether or not furniture requires the raycast to successfully collide with the player to be detected.
- float sneakingPerceptionRadius: Another value of perception Radius in case of specifying a custom radius for detecting sneaking players. This value has yet to be implemented however.
- int fieldOfView: Determines the angle in degrees, for which the furniture can see the player from its focus point.
- float gunshotHearingRadius: Determines how far the furniture can hear gunshots.
- float walkingHearingRadius: Determines how far the furniture can hear player movements. (Not yet implemented)
- float sneakingHearingRadius: Determines how far the furniture can hear the player sneaking. (Not yet implemented)

Chair and Table Scriptable Objects now have placeholder variables for the new sensory radius. Height and xray are unset as it's yet to be determined if these properties are needed.

WanderAI changes:
New Variables:
 - float:
 -- sneakingPerceptionRadius: Not implemented yet. Must first understand how the sneaking system works to integrate it properly.
 -- height: Vertical offset for the transform position when handling sight and sound.
 -- gunshotHearingRadius: Maximum distance to detect gunshot sounds. Any calls for OnGunShoot() further than that distance do nothing.
 -- walkingHearingRadius: Not implemented yet. Needs a sound emitting event of sorts to implement. May be merged with the other 3 hearing radiuses and multiplied by the sound loudness.
 -- sneakingHearingRadius: Not implemented yet. Needs a sound emitting event of sorts to implement. May be merged with the other 3 hearing radiuses and multiplied by the sound loudness.
 -- alertness: Determines the behavior of the AI. Can only be between 0 and 100. At 50 the AI's wandering will interrupt itself ONCE and start wandering at a random spot near the player. At 75 the AI will move towards or away from the player. At 100 the AI will attack the player if it's a predator.
 - int:
 -- fieldOfView: Determines the max angle from the point of focus in which the player can be seen.
 -- investigate: Serves as a resettable switch that allows the sudden change of wander destination to happen ONLY once.
 - bool xray: Determines whether or not the raycast should be ignored. Currently unused until raycast obstacle detection somehow works.
 - GameObject subject: Reference to the player. Is used for updating proximity and sight detection.
Added new OnDrawGizmos() method, which draws the hearing radiuses as wire spheres with specific color coding, and also draws a line between itself and the player, as well as a "cone" wireframe to visualize the sight. Sight outline changes color if the AI can detect the player.
Modified Update() method, wandering behavior remains default below 50 alertness, but otherwise it will attempt to interrupt itself once using investigate to check if it already interrupted itself, and then attempt to wander towards the player, with random offset. a bool detected is also used, and the collider method that determines its behavior has been modified to only set subject (Might be removed later since it's probably not needed, and sets detected to true if player is detected. detected is used to determine whether alertness must tick up or down, and alertness constantly clamps itself between 0 and 100. At 50 alertness investigate is set to 1 if it's set to 0 to initialize the first interruption, at 75 alertness the original pursue/flee behavior activates with the exception of the attack function, at 100 alertness the attack function can occur. Otherwise the default undetected behavior occurs.
OnGunShoot() has been modified to get the player's distance and use gunshotHearingRadius to determine whether or not to add 80 to alertness.

Raycasting has yet to work because of unknown complications that keep the raycast from detecting the player, and sound emitting events have yet to be implemented to start working on the hearing functions to add alertness.