Scriptable Object Sense Implementation
Changes:
PlayerMovement now has an extra bool called isMoving, which is used for "Quiet" and "Normal" sound detection (But not "Loud" sound detection). It also has a function that toggles that bool based on whether or not the player is moving. This may/may not work as it's unknown whether Move() is still called when there's no input.
3 new Scripted Objects have been created:
SenseSO contains two strings called id and name, a Vector3 offset, a Quaternion rotOffset, two colors called debugIdleColor and debugDetectedColor, a SenseType enum called senseType, and a SenseCategory enum called senseCategory.
SenseConeSO is a subclass of SenseSO that contains two floats called range and maxAngle
SenseSphere is a subclass of SenseSO that contains a float called radius
FurnitureSO got a major cleanup, removing all those sensory radiuses and height, replacing it with an array of SenseSO classes.
New sense objects have been created which follow the documentation.
WanderAI also got a minor cleanup removing unnecessary values and functions, all of the functions and portions of code that rely on senses now loops through the entire array of senses. For preview and actual behavior, Player Sneaking and sense category is taken into account when deciding on which senses to use and which senses to ignore. Based on the senseCategory and senseType, each sense will light up in the preview if it detects the player. Behaviorwise, they will set detected to true whilst incrementing the alertness. If detected is false, the alertness will decrement, otherwise it will do nothing, as each individual sense handles its own alertness increment. OnGunShoot also utilizes senses, except it only targets senseCategories of Loud, which also have a senseType of Sound. IsPlayerMoving() has also been added which behaves just like IsPlayerSneaking() but checks isMoving instead.