Reworked hitbox system
Hitboxes can now be on their own gameobjects as long as they remain a child of the parent shootable.
Hitboxes also display their modifier visually with a gradient from 0 (blue) to 1 (green) to 2+ (red)
Overall, this should make it easier to see and organise the individual hitboxes in the editor.