Part 2 of Wander AI Changes + Preparations
Added a few extra values for the Scriptable Objects.
Added a few tags to help with utilizing raycasting.
Cleaned out a few unused functions and variables.
Implemented some of the new variables added in the Scriptable Objects.
Implemented Raycast line of sight for Gizmo Previewing and senses as well.
Separated chunks of code in Update() into separate functions.
Removed a few magic numbers replacing them with the SO values.