Part 3 of Wander AI Changes + Misc
Updated Wander AI:
- Further Expansions to PathfindingState allowing AI to react to players even if they aren't fully alerted.
- Implemented Alert Ability that behaves like SoundAlerter but maxes out alertness instantly for nearby aggressive furniture when fully alerted.
- Updated the UpdateWanderAI function to properly separate the different behaviors based on each alertness threshold. 1st behavior is normal wandering behavior, 2nd behavior is halved speed whilst wandering shorter distance to replicate 'looking around', 3rd behavior is either running PreyAI() to move away from the player, or pathfinding directly to the player, depending on behaviorType. 4th and final behavior is handled by UpdateAlertAI()
- Added a public bool function that determines if the furniture is aggressive or not. Any furniture with the Prey AIType is passive, whereas anything else is aggressive.
- IncrementAlertness now has an optional bool that is false by default, which when set to true allows it to be multiplied further by a factor depending on the alertRate specified
- Since Player Sneaking will now only effect sounds, all interactions and dependencies on player sneaking have been removed.

SoundAlerter now implements soundFalloff that uses a root function to reduce the alertness gain the further away the source is. Later on it might be a good idea to have the sound alertness handling per furniture handled within WanderAI with a specialized function. Why? Because should that be implemented, there's the possibility of having multiple soundFalloff formulae implemented if you want to add furniture that is more sensitive to sound, meaning that the soundFalloff won't easily kick in unless you're near the very edge.

SenseSO has the SenseCategory removed since only sight is handled in this way, and stealth no longer applies to sight.

SenseSO objects relating to stealth have been removed, and default senses have been renamed.

An attempt has been made to fix the guns not alerting any furniture, including adjusting their SO and adjusting the Shoot() function.

KNOWN ISSUES: The guns still don't alert the furniture!