Implemented aiEntityDistance
An attempt on fixing pathfinding delays by reducing the amount of pathfinding requests done by only handling pathfinding requests that are actually needed. Any AI that is too far away won't run. For now the value is 100 which should be enough, but it may change depending on the map, or if an alternate method is required.