Created instanced behaviours
Behaviours are now Instantiated when they are entered and Destroyed on exit. This allows for behaviours to have instance variables. dangerPosition has been made an instance variable in Flee and Search
AI now holds where it's looking for a brief moment during Search
Entry and Exit functions have also been added to allow for changes to always occur on state changes.