Fixed Charge and Slam Behaviour Damage
Removed some portions of code that were supposed to optimize it, which only turned out to break it later on.
Fixed the charge behavior so that it only deals 1 damage instead of one-shotting the player, and adjusted the method so that no errors pop up.
Slam Behaviour also applies this, minus the variable and knockback. The previous optimization has been removed since it doesn't look like it ever collides with the player.